backend/requirements.txt
Django==3.2.7
django-cors-headers==3.8.0
django-filter==2.4.0
djangorestframework==3.12.4
djangorestframework-simplejwt==4.8.0
pytest==6.2.5
pytest-django==4.4.0
psycopg2==2.9.1
